# `vec_ptype()` and `vec_ptype_finalise()` improvements with data frames