--- a/src/proxy-restore.c
+++ b/src/proxy-restore.c
@@ -91,19 +91,19 @@ SEXP vec_restore_default(SEXP x, SEXP to) {
   return x;
 }
 
-static SEXP df_restore_impl(SEXP x, SEXP to, R_len_t size);
+static SEXP bare_df_restore_impl(SEXP x, SEXP to, R_len_t size);
 
-SEXP vctrs_df_restore(SEXP x, SEXP to, SEXP n) {
+SEXP vctrs_bare_df_restore(SEXP x, SEXP to, SEXP n) {
   if (TYPEOF(x) != VECSXP) {
     Rf_errorcall(R_NilValue, "Internal error: Attempt to restore data frame from a %s.",
                  Rf_type2char(TYPEOF(x)));
   }
 
   R_len_t size = (n == R_NilValue) ? df_raw_size(x) : r_int_get(n, 0);
-  return df_restore_impl(x, to, size);
+  return bare_df_restore_impl(x, to, size);
 }
 
-static SEXP df_restore_impl(SEXP x, SEXP to, R_len_t size) {
+static SEXP bare_df_restore_impl(SEXP x, SEXP to, R_len_t size) {
   x = PROTECT(r_maybe_duplicate(x));
   x = PROTECT(vec_restore_default(x, to));
 
@@ -128,12 +128,12 @@ SEXP vec_restore(SEXP x, SEXP to, SEXP n) {
   default: return vec_restore_dispatch(x, to, n);
   case vctrs_class_none: return vec_restore_default(x, to);
   case vctrs_class_bare_data_frame:
-  case vctrs_class_bare_tibble: return vctrs_df_restore(x, to, n);
+  case vctrs_class_bare_tibble: return vctrs_bare_df_restore(x, to, n);
   case vctrs_class_data_frame: {
     // Restore methods are passed the original atomic type back, so we
     // first restore data frames as such before calling the restore
     // method, if any
-    SEXP out = PROTECT(vctrs_df_restore(x, to, n));
+    SEXP out = PROTECT(vctrs_bare_df_restore(x, to, n));
     out = vec_restore_dispatch(out, to, n);
     UNPROTECT(1);
     return out;