--- a/src/utils.c
+++ b/src/utils.c
@@ -256,7 +256,7 @@ SEXP map(SEXP x, SEXP (*fn)(SEXP)) {
 
 SEXP df_map(SEXP df, SEXP (*fn)(SEXP)) {
   SEXP out = PROTECT(map(df, fn));
-  out = vctrs_bare_df_restore(out, df, vctrs_shared_zero_int);
+  out = vec_bare_df_restore(out, df, vctrs_shared_zero_int);
 
   UNPROTECT(1);
   return out;