--- a/src/vctrs.h
+++ b/src/vctrs.h
@@ -389,7 +389,6 @@ R_len_t df_rownames_size(SEXP x);
 R_len_t df_raw_size(SEXP x);
 R_len_t df_raw_size_from_list(SEXP x);
 SEXP vctrs_df_restore(SEXP x, SEXP to, SEXP n);
-SEXP df_restore_impl(SEXP x, SEXP to, R_len_t size);
 
 SEXP chr_assign(SEXP out, SEXP index, SEXP value, bool clone);
 SEXP list_assign(SEXP out, SEXP index, SEXP value, bool clone);