--- a/src/vctrs.h
+++ b/src/vctrs.h
@@ -342,7 +342,7 @@ SEXP vec_assign(SEXP x, SEXP index, SEXP value);
 bool vec_requires_fallback(SEXP x, struct vctrs_proxy_info info);
 SEXP vec_init(SEXP x, R_len_t n);
 SEXP vec_type(SEXP x);
-SEXP vec_type_finalise(SEXP x);
+SEXP vec_ptype_finalise(SEXP x);
 bool vec_is_unspecified(SEXP x);
 SEXP vec_recycle(SEXP x, R_len_t size, struct vctrs_arg* x_arg);
 SEXP vec_recycle_common(SEXP xs, R_len_t size);