--- a/tests/testthat/test-type-tibble.R
+++ b/tests/testthat/test-type-tibble.R
@@ -58,3 +58,17 @@ test_that("the type of a tibble with an unspecified column retains unspecifiedne
   expect_identical(vec_ptype(df1), expect)
   expect_identical(vec_ptype(df2), expect)
 })
+
+test_that("vec_ptype_finalise() works recursively over tibbles", {
+  df <- tibble(x = numeric(), y = unspecified())
+  expect <- tibble(x = numeric(), y = logical())
+
+  expect_identical(vec_ptype_finalise(df), expect)
+})
+
+test_that("vec_ptype_finalise() can handle tibble df columns", {
+  df <- tibble(x = numeric(), y = tibble(z = unspecified()))
+  expect <- tibble(x = numeric(), y = tibble(z = logical()))
+
+  expect_identical(vec_ptype_finalise(df), expect)
+})