Closes #800 

This PR has two goals:

- Rework `vec_type()` to be recursive over all data frame classes, rather than calling `vec_slice(x, NULL)`. This ensures unspecified columns are handled correctly.

- Rework `vec_ptype_finalise()` to be recursive over all data frame classes. This ensures that unspecified columns are finalised to logicals.

Along the way I changed:

- Renamed `vctrs_df_restore()` -> `vec_bare_df_restore()`

- New `vec_df_restore()` that restores generically by dispatching to the R method if one exists

- These two things make `vec_restore()` a bit cleaner to follow

- Similarly, renamed `df_map()` to `bare_df_map()`

- New `df_map()` that restores with the generic `vec_df_restore()`. We use this to map over classed data frames

- `s3_type()` now uses `class_type()` to allow us to recursively map over data frames. Bare tibbles are optimized, and data frames with unknown classes go through the generic map. This fixes the first problem above.

- `vec_ptype_finalise()` has been simplified a bit, and now uses `class_type()` to be recursive over all data frame classes. This fixes the second problem. `vec_ptype_finalise()` also didn't have many tests, so I added a few more.

Mainly this now works correctly:

``` r
library(vctrs)
library(tibble, warn.conflicts = FALSE)

internal_ptype <- vec_ptype(tibble(x = 2, y = NA))
internal_ptype
#> # A tibble: 0 x 2
#> # … with 2 variables: x <dbl>, y <???>

vec_ptype_finalise(internal_ptype)
#> # A tibble: 0 x 2
#> # … with 2 variables: x <dbl>, y <lgl>
```

We get nice performance improvements in determining the type of tibbles. This isn't done directly that often, but is still useful:

```r
library(vctrs)
library(tibble, warn.conflicts = FALSE)
library(rlang)

# 1 row, 100 cols
x <- as_tibble(set_names(rep_len(list(1), 100), paste0("a", 1:100)))
dim(x)
#> [1]   1 100

# before
bench::mark(vec_ptype(x))
#> # A tibble: 1 x 6
#>   expression        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype(x)   17.5µs   19.6µs    48136.    7.15KB     24.1

# after
bench::mark(vec_ptype(x))
#> # A tibble: 1 x 6
#>   expression        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype(x)   2.99µs   3.63µs   253962.    7.15KB     25.4
```